Clean architecture.
