+++ app/src/main/java/ru/brauer/nearspace/data/network/CallbackApodResponse.kt
@@ -1,0 +1,8 @@
+package ru.brauer.nearspace.data.network
+
+import ru.brauer.nearspace.data.network.response.ApodResponse
+
+interface CallbackApodResponse {
+    fun notify(response: ApodResponse?, onServerFailureMessage: String? = null)
+    fun notify(ex: Throwable)
+}
