+++ app/src/main/java/ru/brauer/nearspace/data/repository/CallbackApod.kt
@@ -1,0 +1,8 @@
+package ru.brauer.nearspace.data.repository
+
+import ru.brauer.nearspace.domain.entities.Apod
+
+interface CallbackApod {
+        fun notify(response: Apod?, onServerFailureMessage: String? = null)
+        fun notify(ex: Throwable)
+}
