+++ app/src/main/java/ru/brauer/nearspace/domain/entities/Apod.kt
@@ -1,0 +1,8 @@
+package ru.brauer.nearspace.domain.entities
+
+data class Apod(
+    val mediaType: String,
+    val url: String,
+    val photoDescription: String,
+    val title: String
+)
