+++ app/src/main/java/ru/brauer/nearspace/domain/interactor/ApodInteractor.kt
@@ -1,0 +1,11 @@
+package ru.brauer.nearspace.domain.interactor
+
+import ru.brauer.nearspace.domain.entities.Apod
+
+interface ApodInteractor {
+    fun getApod(date: Long? = null, callbackApod: CallbackApod)
+
+    interface CallbackApod {
+        fun notify(apod: Apod?, onServerFailureMessage: String?, ex: Throwable?)
+    }
+}
