--- app/src/main/java/ru/brauer/nearspace/domain/repository/NasaAPI.kt
@@ -1,14 +1,0 @@
-package ru.brauer.nearspace.domain.repository
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-import ru.brauer.nearspace.domain.repository.dto.ApodDTO
-
-interface NasaAPI {
-    @GET("planetary/apod")
-    fun getApod(
-        @Query("api_key") apiKey: String,
-        @Query("date") date: String?
-    ): Call<ApodDTO>
-}
