+++ app/src/main/java/ru/brauer/nearspace/domain/repository/Repository.kt
@@ -1,8 +1,7 @@
 package ru.brauer.nearspace.domain.repository
 
-import retrofit2.Callback
-import ru.brauer.nearspace.domain.repository.dto.ApodDTO
+import ru.brauer.nearspace.data.repository.CallbackApod
 
 interface Repository {
-    fun getApod(date:String?, callback: Callback<ApodDTO>)
+    fun getApod(date:String?, callback: CallbackApod)
 }
