+++ app/src/main/java/ru/brauer/nearspace/presentation/earth/EarthFragment.kt
@@ -1,0 +1,34 @@
+package ru.brauer.nearspace.presentation.earth
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import ru.brauer.nearspace.databinding.FragmentEarthBinding
+
+class EarthFragment : Fragment() {
+
+    private var _binding: FragmentEarthBinding? = null
+    private val binding get() = _binding!!
+
+    companion object {
+
+        @JvmStatic
+        fun newInstance() =
+            EarthFragment()
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentEarthBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
