+++ app/src/main/java/ru/brauer/nearspace/presentation/mars/MarsFragment.kt
@@ -1,0 +1,33 @@
+package ru.brauer.nearspace.presentation.mars
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import ru.brauer.nearspace.databinding.FragmentMarsBinding
+
+class MarsFragment : Fragment() {
+
+    private var _binding: FragmentMarsBinding? = null
+    private val binding get() = _binding!!
+
+    companion object {
+        @JvmStatic
+        fun newInstance() =
+            MarsFragment()
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentMarsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
