+++ app/src/main/java/ru/brauer/nearspace/presentation/settings/AppTheme.kt
@@ -1,0 +1,8 @@
+package ru.brauer.nearspace.presentation.settings
+
+import ru.brauer.nearspace.R
+
+enum class AppTheme(val key: String, val resource: Int) {
+    DEFAULT("default", R.style.Theme_NearSpace),
+    MARTIAN("martian", R.style.Theme_NearSpace_Martian)
+}
