+++ app/src/main/java/ru/brauer/nearspace/presentation/util/Extensions.kt
@@ -1,0 +1,7 @@
+package ru.brauer.nearspace.presentation.util
+
+import java.text.SimpleDateFormat
+import java.util.*
+
+fun Long.toFormate(pattern: String): String =
+    SimpleDateFormat(pattern, Locale.US).format(Date(this))
