--- app/src/main/java/ru/brauer/nearspace/ui/settings/AppTheme.kt
@@ -1,8 +1,0 @@
-package ru.brauer.nearspace.ui.settings
-
-import ru.brauer.nearspace.R
-
-enum class AppTheme(val key: String, val resource: Int) {
-    DEFAULT("default", R.style.Theme_NearSpace),
-    MARTIAN("martian", R.style.Theme_NearSpace_Martian)
-}
