--- app/src/main/java/ru/brauer/nearspace/ui/weather/WeatherFragment.kt
@@ -1,34 +1,0 @@
-package ru.brauer.nearspace.ui.weather
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import ru.brauer.nearspace.R
-import ru.brauer.nearspace.databinding.FragmentWeatherBinding
-
-class WeatherFragment : Fragment() {
-
-    private var _binding: FragmentWeatherBinding? = null
-    private val binding get() = _binding!!
-
-    companion object {
-        @JvmStatic
-        fun newInstance() =
-            WeatherFragment()
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentWeatherBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
