--- app/src/main/java/ru/brauer/nearspace/util/Extensions.kt
@@ -1,7 +1,0 @@
-package ru.brauer.nearspace.util
-
-import java.text.SimpleDateFormat
-import java.util.*
-
-fun Long.toFormate(pattern: String): String =
-    SimpleDateFormat(pattern, Locale.US).format(Date(this))
