+++ app/src/main/res/layout/activity_main.xml
@@ -23,7 +23,7 @@
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight="1"
-            tools:context=".ui.MainActivity" />
+            tools:context=".presentation.MainActivity" />
 
     </LinearLayout>
