+++ app/src/main/res/layout/fragment_main.xml
@@ -10,7 +10,7 @@
 
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.MainFragment">
+        tools:context=".presentation.main.MainFragment">
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/input_layout"
