+++ app/src/main/res/layout/fragment_mars.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.mars.MarsFragment">
+    tools:context=".presentation.mars.MarsFragment">
 
     <ImageView
         android:layout_width="match_parent"
