+++ app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.settings.SettingsFragment">
+    tools:context=".presentation.settings.SettingsFragment">
 
     <TextView
         android:layout_width="wrap_content"
