+++ app/src/main/res/layout/fragment_weather.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.weather.WeatherFragment">
+    tools:context=".presentation.weather.WeatherFragment">
 
     <ImageView
         android:layout_width="match_parent"
